+++ pages/api/admin/subCategory.js
@@ -1,11 +1,12 @@
-import Category from "../../../models/Category";
-import SubCategory from "../../../models/SubCategory";
+import SubCategory from "@/models/SubCategory";
 import nc from "next-connect";
-import auth from "../../../middleware/auth";
-import admin from "../../../middleware/admin";
-import db from "../../../utils/db";
+import auth from "@/middleware/auth";
+import admin from "@/middleware/admin";
+import db from "@/utils/db";
 import slugify from "slugify";
 import Product from "@/models/Product";
+import GroupSubCategory from "@/models/GroupSubCategory";
+import Category from "@/models/Category";
 
 const handler = nc().use(auth).use(admin);
 
@@ -17,7 +18,7 @@ handler.post(async (req, res) => {
     if (test) {
       return res
         .status(400)
-        .json({ message: "SubCategory already exist, try a different name" });
+        .json({ message: "Така назва підкатегорії вже існує, виберіть нову" });
     }
     await new SubCategory({
       name,
@@ -26,14 +27,17 @@ handler.post(async (req, res) => {
       slug: slugify(name, "_"),
     }).save();
     await db.disconnectDb();
-    res.json({
-      message: `SubCategory ${name} has been created successfully`,
-      subCategories: await SubCategory.find({}).sort({ updatedAt: -1 }),
+   return res.json({
+      message: `Підкатегорія ${name} створена успішно`,
+      subCategories: await SubCategory.find({})
+      .populate({ path: "parent", model: GroupSubCategory })
+      .populate({ path: "top_parent", model: Category })
+      .sort({ name: 1 }),
     });
   } catch (error) {
     await db.disconnectDb();
     console.log("31", error.message);
-    res.status(500).json({ message: error.message });
+   return res.status(500).json({ message: error.message });
   }
 });
 
@@ -41,20 +45,35 @@ handler.delete(async (req, res) => {
   try {
     const { id } = req.body;
     await db.connectDb();
-    const productsToDelete = await Product.find({ subCategories: { $in: [id] } });
-    let countDeletedProduct = 0;
-    for (const element of productsToDelete) {
-      await Product.findByIdAndDelete(element._id);
-      countDeletedProduct++;
+    const existProducts = await Product.find({ subCategories: { $in: [id] } });;
+    if(existProducts.length == 0) {
+      await SubCategory.findByIdAndRemove(id);
+      await db.disconnectDb();
+      return res.json({
+        message: "Підкатегорія успішно видалена!",
+        subCategories: await SubCategory
+        .find({})
+        .populate({ path: "parent", model: GroupSubCategory })
+        .populate({ path: "top_parent", model: Category })
+        .sort({ name: 1 }),
+      });
+    } else {
+      await db.disconnectDb();
+      throw new Error(`У підкатегорії є ${existProducts.length} продукт(-ів), спершу видаліть їх!`)
     }
-    await SubCategory.findByIdAndRemove(id);
-    await db.disconnectDb();
-    return res.json({
-      message: `SubCategory has been deleted succesfuly and ${countDeletedProduct} products has been deleted too`,
-      subCategories: await SubCategory.find({}).sort({ updatedAt: -1 }),
-    });
+    // let countDeletedProduct = 0;
+    // for (const element of productsToDelete) {
+    //   await Product.findByIdAndDelete(element._id);
+    //   countDeletedProduct++;
+    // }
+    // await SubCategory.findByIdAndRemove(id);
+    // await db.disconnectDb();
+    // return res.json({
+    //   message: `SubCategory has been deleted succesfuly and ${countDeletedProduct} products has been deleted too`,
+    //   subCategories: await SubCategory.find({}).sort({ updatedAt: -1 }),
+    // });
   } catch (error) {
-    res.status(500).json({ message: error.message });
+   return res.status(500).json({ message: error.message });
   }
 });
 
@@ -70,11 +89,14 @@ handler.put(async (req, res) => {
     });
     await db.disconnectDb();
     return res.json({
-      message: "SubCategory has been updated succesfuly",
-      subCategories: await SubCategory.find({}).sort({ updatedAt: -1 }),
+      message: "Підкатегорія успішно змінена",
+      subCategories: await SubCategory.find({})
+      .populate({ path: "parent", model: GroupSubCategory })
+      .populate({ path: "top_parent", model: Category })
+      .sort({ name: 1 }),
     });
   } catch (error) {
-    res.status(500).json({ message: error.message });
+   return res.status(500).json({ message: error.message });
   }
 });
 
@@ -87,11 +109,13 @@ handler.get(async (req, res) => {
     await db.connectDb();
     const result = await SubCategory.find({ parent: groupSubCategory }).select(
       "name"
-    );
+    ).sort({
+      name: 1,
+    });
     await db.disconnectDb();
     return res.json(result);
   } catch (error) {
-    res.status(500).json({ message: error.message });
+   return res.status(500).json({ message: error.message });
   }
 });
